refactor(daemon): extract boot_recovery from lib.rs (#276)
Behavior-preserving carve of the daemon boot-recovery concern (#213/#215)
out of lib.rs into a new boot_recovery.rs, continuing the #276 module
decomposition (after worktree_ops #276-Slice1, run_advance #235, and the
event_log::project #238 precedent).

Moved verbatim into boot_recovery.rs: run_boot_recovery (now pub(crate),
the sole widening) plus the three private helpers detect_merged_without_event,
merged_without_event_nodes, and branch_is_merged_into. Repointed all four
call sites (the run_boot_recovery_tick seam, the serve_with_config startup
pass, and two in-test calls) to boot_recovery::run_boot_recovery.

Boundaries honored: reconcile_run_level_stall and find_node_type stay in
lib.rs (shared with the stale sweep) and are called up into via crate::;
run_state_with_node is duplicated (not moved) since the stall tests still
need the lib.rs copy; the deliberate non-is_terminal() set (#237 follow-up
F1) is moved verbatim, not "fixed". No CONTEXT.md/ADR churn.

Tests: the 3 closure-injected pure tests move into boot_recovery::tests
alongside a new git-backed branch_is_merged_into_tracks_ancestry test
exercising the real `git merge-base --is-ancestor` path; the AppState-backed
boot_recovery_reconciles_session_holding_node_in_terminal_run stays in
crate::tests (repointed). cargo test -p pdo-daemon: 1207 passed.
